+++ tempodb/tempodb_search_test.go
@@ -103,7 +103,13 @@ func advancedTraceQLRunner(t *testing.T, wantTr *tempopb.Trace, wantMeta *tempop
 	falseConditions := []string{
 		fmt.Sprintf("name=`%v`", test.RandomString()),
 		fmt.Sprintf("duration>%dh", rand.Intn(10)+1),
+		fmt.Sprintf("rootServiceName=`%v`", test.RandomString()),
 		// status? can't really construct a status condition that's false for all spans
+	}
+	trueTraceC := []string{
+		fmt.Sprintf("traceDuration=%dms", wantMeta.DurationMs),
+		fmt.Sprintf("rootServiceName=`%s`", wantMeta.RootServiceName),
+		fmt.Sprintf("rootName=`%s`", wantMeta.RootTraceName),
 	}
 	totalSpans := 0
 	for _, b := range wantTr.Batches {
@@ -122,6 +128,7 @@ func advancedTraceQLRunner(t *testing.T, wantTr *tempopb.Trace, wantMeta *tempop
 				trueC = append(trueC, fmt.Sprintf("status=%s", status))
 				trueC = append(trueC, fmt.Sprintf("kind=%s", kind))
 				trueC = append(trueC, trueResourceC...)
+				trueC = append(trueC, trueTraceC...)
 
 				trueConditionsBySpan = append(trueConditionsBySpan, trueC)
 				falseConditions = append(falseConditions, falseC...)
@@ -157,7 +164,7 @@ func advancedTraceQLRunner(t *testing.T, wantTr *tempopb.Trace, wantMeta *tempop
 		{Query: fmt.Sprintf("{} | count() != %d", totalSpans+1)},
 		{Query: fmt.Sprintf("{ true } && { true } | count() = %d", totalSpans)},
 		{Query: fmt.Sprintf("{ true } || { true } | count() = %d", totalSpans)},
-		{Query: fmt.Sprintf("{ %s && %s } | count() = 1", rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]))},
+		{Query: fmt.Sprintf("{ %s && %s && name=`MySpan` } | count() = 1", rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]))},
 		// avgs/min/max/sum
 		{Query: fmt.Sprintf("{ %s && %s } && { %s && %s } | avg(duration) = %dns",
 			rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]),
@@ -371,6 +378,31 @@ func groupTraceQLRunner(t *testing.T, wantTr *tempopb.Trace, wantMeta *tempopb.T
 	}
 }
 
+// oneQueryRunner is a good place to place a single query for debugging
+// func oneQueryRunner(t *testing.T, wantTr *tempopb.Trace, wantMeta *tempopb.TraceSearchMetadata, _, _ []*tempopb.SearchRequest, meta *backend.BlockMeta, r Reader) {
+// 	ctx := context.Background()
+// 	e := traceql.NewEngine()
+
+// 	searchesThatMatch := []*tempopb.SearchRequest{
+// 		// conditions
+// 		{Query: "{rootServiceName=`fotlVYVqts`} || {.k8s.container.name=`k8sContainer`}"},
+// 	}
+
+// 	for _, req := range searchesThatMatch {
+// 		fetcher := traceql.NewSpansetFetcherWrapper(func(ctx context.Context, req traceql.FetchSpansRequest) (traceql.FetchSpansResponse, error) {
+// 			return r.Fetch(ctx, meta, req, common.DefaultSearchOptions())
+// 		})
+
+// 		res, err := e.ExecuteSearch(ctx, req, fetcher)
+// 		require.NoError(t, err, "search request: %+v", req)
+// 		actual := actualForExpectedMeta(wantMeta, res)
+// 		require.NotNil(t, actual, "search request: %v", req)
+// 		actual.SpanSet = nil // todo: add the matching spansets to wantmeta
+// 		actual.SpanSets = nil
+// 		require.Equal(t, wantMeta, actual, "search request: %v", req)
+// 	}
+// }
+
 func conditionsForAttributes(atts []*v1_common.KeyValue, scope string) ([]string, []string) {
 	trueConditions := []string{}
 	falseConditions := []string{}
@@ -466,7 +498,7 @@ func runCompleteBlockSearchTest(t *testing.T, blockVersion string, runners ...ru
 	dec := model.MustNewSegmentDecoder(model.CurrentEncoding)
 
 	totalTraces := 50
-	wantTrIdx := rand.Intn(50)
+	wantTrIdx := rand.Intn(totalTraces)
 	for i := 0; i < totalTraces; i++ {
 		var tr *tempopb.Trace
 		var id []byte